pptxreader/requirements.txt
Cython==3.0.5
lxml==4.9.3
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
PyPDF2==3.0.1
python-magic==0.4.27
python-pptx==0.6.23
XlsxWriter==3.1.9
